+++ packages/react/src/contexts/__tests__/ClerkProvider.test.tsx
@@ -16,10 +16,30 @@ import {
   ukUA,
 } from '@clerk/localizations';
 import { dark } from '@clerk/ui/themes';
-import { describe, expectTypeOf, it } from 'vitest';
+import { render } from '@testing-library/react';
+import React from 'react';
+import { afterAll, beforeAll, describe, expect, expectTypeOf, it, vi } from 'vitest';
 
 import type { ClerkProviderProps as GenericClerkProviderProps, Ui } from '../../types';
-import type { ClerkProvider } from '../ClerkProvider';
+import { ClerkProvider } from '../ClerkProvider';
+
+vi.mock('../../isomorphicClerk', () => {
+  let instance: any;
+  class IsomorphicClerk {
+    status = 'loading';
+    on = vi.fn();
+    off = vi.fn();
+    __internal_updateProps = vi.fn().mockResolvedValue(undefined);
+    static getOrCreateInstance() {
+      instance ??= new IsomorphicClerk();
+      return instance;
+    }
+    static clearInstance() {
+      instance = undefined;
+    }
+  }
+  return { IsomorphicClerk };
+});
 
 type ClerkProviderProps = Parameters<typeof ClerkProvider>[0];
 type CustomAppearance = {
@@ -230,6 +250,47 @@ describe('ClerkProvider', () => {
         routerPush: () => {},
         routerReplace: () => {},
       }).toMatchTypeOf<ClerkProviderProps>();
+    });
+  });
+
+  describe('duplicate detection', () => {
+    const pk = 'pk_test_Y2xlcmsuY2xlcmsuZGV2JA';
+    const originalError = console.error;
+
+    beforeAll(() => {
+      console.error = vi.fn();
+    });
+
+    afterAll(() => {
+      console.error = originalError;
+    });
+
+    it('throws when a ClerkProvider is nested inside another ClerkProvider', () => {
+      expect(() =>
+        render(
+          <ClerkProvider publishableKey={pk}>
+            <ClerkProvider publishableKey={pk}>
+              <div />
+            </ClerkProvider>
+          </ClerkProvider>,
+        ),
+      ).toThrow(/multiple <ClerkProvider>/);
+    });
+
+    it('does not throw when a second React root mounts while the first is still mounted', () => {
+      const first = render(
+        <ClerkProvider publishableKey={pk}>
+          <div />
+        </ClerkProvider>,
+      );
+      expect(() =>
+        render(
+          <ClerkProvider publishableKey={pk}>
+            <div />
+          </ClerkProvider>,
+        ),
+      ).not.toThrow();
+      first.unmount();
     });
   });
 });
